Print result of TlsStream::read
